created files and separated out ApiV1 into separate files. Just want to create some docs per file to give an overview of what's happening in each
